+++ arch/arm/src/lc823450/lc823450_timer.c
@@ -727,6 +727,6 @@ int up_rtc_gettime(struct timespec *tp)
   tp->tv_sec  = secs;
   tp->tv_nsec = nsecs;
 
-  tmrinfo("Returning tp=(%d,%d)\n", (int)tp->tv_sec, (int)tp->tv_nsec);
+  tmrinfo("Returning tp=(%jd,%ld)\n", (intmax_t)tp->tv_sec, tp->tv_nsec);
   return OK;
 }
