+++ arch/arm/src/xmc4/xmc4_tickless.c
@@ -375,8 +375,8 @@ int up_timer_gettime(struct timespec *ts)
   ts->tv_sec = sec;
   ts->tv_nsec = (usec - (sec * USEC_PER_SEC)) * NSEC_PER_USEC;
 
-  tmrinfo("usec=%llu ts=(%lu, %lu)\n",
-          usec, (unsigned long)ts->tv_sec, (unsigned long)ts->tv_nsec);
+  tmrinfo("usec=%llu ts=(%jd, %ld)\n",
+          usec, (intmax_t)ts->tv_sec, ts->tv_nsec);
 
   return OK;
 }
@@ -412,8 +412,8 @@ int up_timer_start(const struct timespec *ts)
   uint64_t period;
   irqstate_t flags;
 
-  tmrinfo("ts=(%lu, %lu)\n",
-          (unsigned long)ts->tv_sec, (unsigned long)ts->tv_nsec);
+  tmrinfo("ts=(%jd, %ld)\n",
+          (intmax_t)ts->tv_sec, ts->tv_nsec);
   DEBUGASSERT(ts);
 
   /* Was an interval already running? */
@@ -566,8 +566,8 @@ int up_timer_cancel(struct timespec *ts)
       ts->tv_sec  = sec;
       ts->tv_nsec = nsec;
 
-      tmrinfo("remaining count : %lu (%lu, %lu)\n", count,
-              (unsigned long)ts->tv_sec, (unsigned long)ts->tv_nsec);
+      tmrinfo("remaining count : %lu (%jd, %ld)\n", count,
+              (intmax_t)ts->tv_sec, ts->tv_nsec);
     }
 
   return OK;
